ICAST インタフェース用ダウンキャスト OCAST インタフェース→オブジェクト用 キャスト を導入．
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@886 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDConvexMesh.cpp
@@ -212,7 +212,7 @@ Vec3f CDConvexMesh::Support(const Vec3f& v) const {
 	return Vertex(curPos);
 }
 
-CDFaceIf* CDConvexMesh::GetFace(int i){
+CDFaceIf* CDConvexMesh::GetFace(size_t i){
 	return &faces.at(i);
 }
 size_t CDConvexMesh::GetNFaces(){
